Рефакторит модуль с утилитами
- разбивает утилитарный модуль на несколько модулей
- метод render сразу вызывает getElement компонентов
- добавляет метод replace
