+++ src/main.js
@@ -9,7 +9,8 @@ import DayView from "./view/day.js";
 import EventFormView from "./view/event-form.js";
 import EventPointView from "./view/event-point.js";
 import {generateEventPoint} from "./mock/event-point.js";
-import {render, RenderPosition, getUniqueDates} from "./utils.js";
+import {render, RenderPosition, replace} from "./utils/render.js";
+import {getUniqueDates} from "./utils/specific.js";
 
 const EVENTS_COUNT = 25;
 
@@ -21,38 +22,38 @@ const sortEvents = events.sort((a, b) => a.time.start - b.time.start);
 const siteHeaderElement = document.querySelector(`.page-header`);
 const tripMainElement = siteHeaderElement.querySelector(`.trip-main`);
 
-render(tripMainElement, new TripInfoView(sortEvents).getElement(), AFTERBEGIN);
+render(tripMainElement, new TripInfoView(sortEvents), AFTERBEGIN);
 
 const tripInfoElement = siteHeaderElement.querySelector(`.trip-info`);
 
-render(tripInfoElement, new TripPriceView(sortEvents).getElement(), BEFOREEND);
+render(tripInfoElement, new TripPriceView(sortEvents), BEFOREEND);
 
 const tripControlsElement = siteHeaderElement.querySelector(`.trip-controls`);
 
 render(
     tripControlsElement,
-    new TitleH2View(new SiteMenuView().getTitle()).getElement(),
+    new TitleH2View(new SiteMenuView().getTitle()),
     BEFOREEND
 );
-render(tripControlsElement, new SiteMenuView().getElement(), BEFOREEND);
+render(tripControlsElement, new SiteMenuView(), BEFOREEND);
 
 render(
     tripControlsElement,
-    new TitleH2View(new FilterView().getTitle()).getElement(),
+    new TitleH2View(new FilterView().getTitle()),
     BEFOREEND
 );
-render(tripControlsElement, new FilterView().getElement(), BEFOREEND);
+render(tripControlsElement, new FilterView(), BEFOREEND);
 
 const mainElement = document.querySelector(`.page-main`);
 const tripEventsElement = mainElement.querySelector(`.trip-events`);
 
-render(tripEventsElement, new SortView().getElement(), BEFOREEND);
-render(tripEventsElement, new TripDaysView().getElement(), BEFOREEND);
+render(tripEventsElement, new SortView(), BEFOREEND);
+render(tripEventsElement, new TripDaysView(), BEFOREEND);
 
 const tripDaysElement = mainElement.querySelector(`.trip-days`);
 
 getUniqueDates(events).forEach((item, index) => {
-  render(tripDaysElement, new DayView(item, index).getElement(), BEFOREEND);
+  render(tripDaysElement, new DayView(item, index), BEFOREEND);
 });
 
 const renderEvent = (eventsListElement, event) => {
@@ -60,11 +61,11 @@ const renderEvent = (eventsListElement, event) => {
   const eventEditComponent = new EventFormView(event);
 
   const replaceEventToForm = () => {
-    eventsListElement.replaceChild(eventEditComponent.getElement(), eventComponent.getElement());
+    replace(eventEditComponent, eventComponent);
   };
 
   const replaceFormToEvent = () => {
-    eventsListElement.replaceChild(eventComponent.getElement(), eventEditComponent.getElement());
+    replace(eventComponent, eventEditComponent);
   };
 
   eventComponent.setEditClickHandler(() => {
@@ -75,7 +76,7 @@ const renderEvent = (eventsListElement, event) => {
     replaceFormToEvent();
   });
 
-  render(eventsListElement, eventComponent.getElement(), RenderPosition.BEFOREEND);
+  render(eventsListElement, eventComponent, RenderPosition.BEFOREEND);
 };
 
 for (let i = 0; i < EVENTS_COUNT; i++) {
